refactor(communities): Improving code organization on Communities
