--- src/modules/communities/pages/CommunityPage/CommunitySelectionPage/CommunitySelection.styles.ts
@@ -1,16 +1,0 @@
-import { makeStyles } from '@material-ui/core';
-
-export default makeStyles(theme => ({
-  pageTitle: {
-    marginBottom: 40,
-    [theme.breakpoints.down('sm')]: {
-      marginBottom: 10,
-    },
-  },
-  card: {
-    width: '100%',
-  },
-  emptyStateContainer: {
-    height: '80vh',
-  },
-}));
